cosmos/postprocess/requirements.txt
pymongo
numpy
pillow
pyyaml
hyperyaml
xgboost
click
joblib
Pillow
tqdm
matplotlib
pandas
sklearn
bs4
